build: resolve merge
