--- core/common/src/main/java/com/into/websoso/core/common/lifecycle/compose/FlowExtensions.kt
@@ -1,31 +1,0 @@
-package com.into.websoso.core.common.lifecycle.compose
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.compose.LocalLifecycleOwner
-import androidx.lifecycle.repeatOnLifecycle
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.withContext
-import kotlin.coroutines.CoroutineContext
-import kotlin.coroutines.EmptyCoroutineContext
-
-@Composable
-fun <T> Flow<T>.collectAsEventWithLifecycle(
-    lifecycle: Lifecycle = LocalLifecycleOwner.current.lifecycle,
-    minActiveState: Lifecycle.State = Lifecycle.State.STARTED,
-    context: CoroutineContext = EmptyCoroutineContext,
-    onEvent: suspend (T) -> Unit,
-) {
-    LaunchedEffect(this, lifecycle, minActiveState, context) {
-        lifecycle.repeatOnLifecycle(minActiveState) {
-            if (context == EmptyCoroutineContext) {
-                this@collectAsEventWithLifecycle.collect { onEvent(it) }
-            } else {
-                withContext(context) {
-                    this@collectAsEventWithLifecycle.collect { onEvent(it) }
-                }
-            }
-        }
-    }
-}
